refactor data loader structure
