+++ nf2/evaluation/spherical/b_diff.py
