+++ nf2/evaluation/spherical/b_diff_synoptic.py
