+++ nf2/evaluation/spherical/energy_gradient.py
