+++ nf2/evaluation/spherical/free_energy.py
