+++ nf2/evaluation/spherical/open_fields.py
