+++ nf2/evaluation/spherical/potential_vtk.py
